test: ReservationDaoTest 추가
- 테스트하기 어려웠던 Storage 테스트를 Dao 테스트로 변경
